requirements.txt
anytree==2.8.0
colorlog==6.6.0
fuzzywuzzy==0.18.0
jsonref==0.2
matplotlib==3.5.1
nltk==3.7
numpy==1.22.1
PyYAML==6.0
requests==2.27.1

xeger~=0.3.5